requirements.txt
absl-py==2.1.0
accelerate==0.26.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-cors==0.7.0
aiosignal==1.3.2
airportsdata==20241001
albucore==0.0.23
albumentations==2.0.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrdict==2.0.1
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
blake3==1.0.4
bleach==6.2.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
clip==0.2.0
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
comm==0.2.2
compressed-tensors==0.9.1
contourpy==1.3.1
controlnet-aux==0.0.3
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
datasets==2.20.0
debugpy==1.8.11
decorator==5.1.1
deepspeed==0.16.2
defusedxml==0.7.1
depyf==0.18.0
diffusers==0.27.2
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.13
fastapi==0.108.0
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.17.0
fire==0.7.0
flash-attn
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.5.0
ftfy==6.3.1
future==1.0.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.67.0
gradio==4.21.0
gradio_client==0.12.0
grpcio==1.69.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.25.2
idna==3.10
image-reward==1.5
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
interegular==0.3.3
IOPaint==1.5.4
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
jaxtyping==0.2.36
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jmespath==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lark==1.2.2
lazy_loader==0.4
lightning==2.2.3
lightning-utilities==0.11.2
lm-format-enforcer==0.10.9
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.3
mistune==3.1.0
mmengine==0.10.5
model-index==0.1.11
modelscope==1.22.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.29.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open-clip-torch==2.24.0
openai==1.62.0
openai-clip==1.0.1
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
orjson==3.10.15
oss2==2.17.0
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.7.1
pexpect==4.9.0
piexif==1.1.3
Pillow==9.5.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==18.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycocotools==2.0.8
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-json-logger==3.2.1
python-multipart==0.0.20
python-socketio==5.7.2
pytorch-lightning==1.5.0
pytz==2023.4
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.42.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.22.3
rsa==4.9
ruff==0.9.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
simple-websocket==1.1.0
simsimd==6.2.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.32.0.post1
stringzilla==3.11.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.1.10
tiktoken==0.8.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.12.0
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.6.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
triton==3.1.0
typer==0.15.2
typer-config==1.4.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.2
vllm==0.7.2
wandb==0.19.1
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
